develop

+++ app/lib/db/migration.php
@@ -8,11 +8,10 @@ require_once LIB . 'db/db.php';
 function migrate(string $direction = 'up'): void
 {
     if (!db_table_exists('migrations')) {
-        db('CREATE TABLE migrations (migration VARCHAR(255) PRIMARY KEY, created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP)');
+        db('CREATE TABLE migrations (migration VARCHAR(255) PRIMARY KEY, created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP, batch INT DEFAULT 1)');
     }
-
     $migrations = glob(SRC . 'migrations/*.php');
-
+    $batch = db_value('SELECT MAX(batch) FROM migrations') + 1;
     foreach ($migrations as $migration) {
         ['up' => $up, 'down' => $down] = require_once $migration;
         $migrationName = basename($migration);
@@ -20,15 +19,23 @@ function migrate(string $direction = 'up'): void
             case 'up':
                 if (!db_exists('SELECT * FROM migrations WHERE migration = ?', $migrationName)) {
                     echo "Migrating: $migrationName...\n";
-                    $up();
-                    db('INSERT INTO migrations (migration) VALUES (?)', $migrationName);
+                    if ($up instanceof Closure) {
+                        $up();
+                    } else {
+                        db($up);
+                    }
+                    db('INSERT INTO migrations (migration, batch) VALUES (?, ?)', $migrationName, $batch);
                     echo "Migrated: $migrationName\n\n";
                 }
                 break;
             case 'down':
-                if (db_exists('SELECT * FROM migrations WHERE migration = ?', $migrationName)) {
+                if (db_exists('SELECT * FROM migrations WHERE migration = ? and batch = (SELECT MAX(batch) FROM migrations)', $migrationName)){
                     echo "Rolling back: $migrationName...\n";
-                    $down();
+                    if ($down instanceof Closure) {
+                        $down();
+                    } else {
+                        db($down);
+                    }
                     db('DELETE FROM migrations WHERE migration = ?', $migrationName);
                     echo "Rolled back: $migrationName\n\n";
                 }
@@ -54,8 +61,18 @@ function make_migration(string $name): void
         throw new Exception('Please provide a name for the migration');
     }
 
-    $migration = date('YmdHis') . '_' . $name . '.php';
+    $migration = date('Ymd_His') . '_' . $name . '.php';
     $path = SRC . 'migrations/' . $migration;
+    preg_match('/^create_(.*)_table/', $name, $matches);
+
+    if ($matches) {
+        $table = $matches[1];
+        $up = "CREATE TABLE $table (id INT AUTO_INCREMENT PRIMARY KEY)";
+        $down = "DROP TABLE $table";
+    } else {
+        $up = $down = '';
+    }
+
     file_put_contents($path,
 <<<PHP
     <?php
@@ -63,12 +80,19 @@ function make_migration(string $name): void
         return [
             'up' => function() {
                 // Write your migration code here
+                db('{$up}');
             },
             'down' => function() {
                 // Write your migration code here
+                db('{$down}');
             }
         ];
 PHP);
 
-    echo "Created migration: $migration\n";
+    echo "Created migration: $path\n";
+}
+
+function migration_get_files(): array
+{
+    return glob(SRC . 'migrations/*.php');
 }
